Fix invalid value being set for the 'target' compiler option
In tsconfig.json the 'target' compiler option is specified as a string
("ES5", "ES6" etc.) however createProgram() expects that option to be
an integer (actually an enum in TypeScript). Previously the 'target'
option was passed through to createProgram() as a string and this was
causing unexpected failures in some cases, this has now been fixed by
validating and converting the compiler options from the tsconfig.json.